+++ CHANGELOG.md
@@ -4,4 +4,4 @@ CHANGELOG
 This changelog references the relevant changes (bug and security fixes) done.
 
 * 0.1.0 (2013-01-31)
- * START: Add following componemt kernel, config, system, dependency
+ * 11db33e: Add following componemt kernel, config, system, dependency